refactor(emails): consolidate into src/lib/emails (CP-2)
Resolve debitos 8 e 9: src/emails e src/lib/email.tsx tinham

responsabilidades sobrepostas e o sender file de 520 linhas

concentrava transporter + 19 senders.

Estrutura nova:

- src/lib/emails/mailer.ts: transporter, sendEmail, sendBestEffort

- src/lib/emails/senders/auth.tsx: 7 senders auth

- src/lib/emails/senders/payments.tsx: 10 senders payment

- src/lib/emails/senders/admin.tsx: 1 sender admin

- src/lib/emails/senders/contact.tsx: 1 sender contact

- src/lib/emails/{components,templates,render,constants,__tests__}: moved

from src/emails/ via git mv (history preserved).

Consumers atualizados: 6 services + lib/auth.ts + lib/auth/hooks.ts

+ lib/email-dispatcher.ts + 2 test files (mock.module path + spyOn).

Deletado: src/lib/email.tsx, src/emails/.

Validacao: 1425 tests pass em escopo amplo (auth + admin + payments

+ public + lib + plugins). 2 fails sao pre-existentes (subscription

TrialPlanMisconfiguredError + orphaned pagarme cleanup), confirmados

via stash em preview.
