+++ src/lib/email-dispatcher.ts
@@ -1,4 +1,4 @@
-import { sendWelcomeEmail } from "@/lib/email";
+import { sendWelcomeEmail } from "@/lib/emails/senders/auth";
 
 export const EmailDispatcher = {
   sendWelcomeEmail,
