+++ src/lib/emails/components/email-info-table.tsx
