+++ src/lib/emails/constants.ts
