+++ src/lib/emails/templates/auth/account-activation.tsx
