+++ src/lib/emails/templates/auth/organization-invitation.tsx
