+++ src/lib/emails/templates/auth/welcome.tsx
