+++ src/modules/payments/admin-provision/admin-provision.service.ts
@@ -4,7 +4,8 @@ import type { AdminOrgProvision } from "@/db/schema";
 import { schema } from "@/db/schema";
 import { env } from "@/env";
 import { auth } from "@/lib/auth";
-import { sendBestEffort, sendProvisionCheckoutLinkEmail } from "@/lib/email";
+import { sendBestEffort } from "@/lib/emails/mailer";
+import { sendProvisionCheckoutLinkEmail } from "@/lib/emails/senders/payments";
 import { logger } from "@/lib/logger";
 import { AdminCheckoutService } from "@/modules/payments/admin-checkout/admin-checkout.service";
 import { PlansService } from "@/modules/payments/plans/plans.service";
