+++ src/modules/payments/checkout/checkout.service.ts
@@ -1,7 +1,7 @@
 import { eq } from "drizzle-orm";
 import { db } from "@/db";
 import { schema } from "@/db/schema";
-import { sendCheckoutLinkEmail } from "@/lib/email";
+import { sendCheckoutLinkEmail } from "@/lib/emails/senders/payments";
 import { Retry } from "@/lib/utils/retry";
 import { CustomerService } from "@/modules/payments/customer/customer.service";
 import { EmailNotVerifiedError } from "@/modules/payments/errors";
